Move settings to mixin classes
